updated setup to uv, removed pixi updated pre-commit
