+++ .gitignore
@@ -122,6 +122,7 @@ celerybeat.pid
 # Environments
 .env
 .venv
+.venv*
 env/
 venv/
 ENV/
@@ -167,3 +168,5 @@ _secret
 .idea
 ,pixi
 config_local.yaml
+
+_data
