--- pixi.toml
@@ -1,60 +1,0 @@
-[project]
-name = "PyTube"
-version = "0.9.2"
-description = "Add a short description here"
-authors = ["Alexander CS Hendorf <alexander@pioneershub.org>"]
-channels = ["conda-forge"]
-platforms = ["osx-arm64", "win-64", "linux-64", "osx-64"]
-
-[tasks]
-postinstall = "pip install --no-build-isolation --no-deps --disable-pip-version-check -e ."
-
-[dependencies]
-python = ">=3.10,<3.13"
-pixi-pycharm = ">=0.0.6,<0.0.7"
-openpyxl = ">=3.1.4,<3.2"
-omegaconf = ">=2.3.0,<2.4"
-
-[pypi-dependencies]
-pytanis = "*"
-pyvimeo = "*"
-colorama = "*"
-structlog = "*"
-rich = "*"
-google-api-python-client = "*"
-oauth2client = "*"
-jinja2 = "*"
-openai = ">=1.47.0, <2"
-mkdocs-material = { version = ">=9.5.39, <10", extras = ["imaging"] }
-mkdocs = ">=1.6.1, <2"
-bumpversion = ">=0.6.0, <0.7"
-pymdown-extensions = ">=10.11.2, <11"
-mkdocs-glightbox = ">=0.4.0, <0.5"
-mkdocs-open-in-new-tab = ">=1.0.5, <2"
-
-[tool.pixi.dependencies]
-ruff = ">=0.4.8,<0.5"
-
-[feature.lint.dependencies]
-pre-commit = "*"
-[feature.lint.tasks]
-lint = "pre-commit run --all"
-
-[feature.test.dependencies]
-pytest = "*"
-pytest-md = "*"
-pytest-mock = ">=3.14.0, <4"
-pytest-emoji = "*"
-hypothesis = "*"
-pytest-cov = "*"
-[feature.test.tasks]
-test = "pytest"
-
-
-[environments]
-default = ["test"]
-lint = ["lint"]
-py312 = ["py312", "test"]
-
-[feature.py312.dependencies]
-python = "3.12.*"
