+++ pyproject.toml
@@ -1,48 +1,118 @@
 [project]
-name = "Vimeo to YouTube"
+name = "video_manager"
 version = "0.9.2"
 description = "Download videos from Vimeo to publish on YouTube, add metadata from pretalx."
-authors = [
-    { name = "Alexander CS Hendorf", email = "hendorf@gmail.com" }
-]
+authors = [{ name = "Alexander CS Hendorf", email = "alexander@pioneershub.org" }]
+
 readme = "README.md"
-license = { file = "LICENSE" }
-classifiers = [
-    "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.12",
-    "Programming Language :: Python :: 3.11",
-    "Programming Language :: Python :: 3.10",
+license-files = [
+    "LICENSE",
+]
+requires-python = ">=3.11,<=3.13"
+
+dependencies = [
+    "pytanis>=0.8",
+    "colorama>=0.4.6",
+    "rich>=14.0.0",
+    "structlog>=25.3.0",
+    "pyvimeo>=1.1.2",
+    "google-api-python-client>=2.169.0",
+    "oauth2client>=4.1.3",
+    "jinja2>=3.1.6",
+    "openai>=1.78.0",
+    "omegaconf>=2.3.0",
+    "pytest>=7.4.4",
 ]
 
-[build-system]
-requires = ["setuptools"]
-build-backend = "setuptools.build_meta"
+[project.optional-dependencies]
+docs = [
+    "mkdocs",
+    "mkdocs-glightbox>=0.4.0",
+    "mkdocs-material[imaging]>=9.6.12",
+    "mkdocs-open-in-new-tab>=1.0.8",
+    "pymdown-extensions>=10.15",
+]
+video_processor = [
+    "numpy>=2.2.5",
+    "opencv-python>=4.11.0.86",
+    "pyyaml>=6.0.2",
+    "dotenv>=0.9.9",
+    "polars>=1.29.0",
+    "fastexcel>=0.14.0",
+    "xlsxwriter>=3.2.3",
+]
+tests = [
+    'pytest >=7, < 8',
+]
 
-[project.urls]
-Homepage = "https://github.com/PioneersHub/vimeo2youtube"
+# Alle Abhängigkeiten für Entwickler
+all = [
+    "video_manager[dev]",
+    "video_manager[docs]",
+    "video_manager[video_processor]",
+]
+
+
+[tool.pytest.ini_options]
+minversion = "7.0"
+addopts = "-ra -q -s"
+testpaths = [
+    "tests",
+]
+exclude = [
+    ".bzr",
+    ".direnv",
+    ".eggs",
+    ".git",
+    ".git-rewrite",
+    ".hg",
+    ".ipynb_checkpoints",
+    ".mypy_cache",
+    ".nox",
+    ".pants.d",
+    ".pyenv",
+    ".pytest_cache",
+    ".pytype",
+    ".ruff_cache",
+    ".svn",
+    ".tox",
+    ".venv",
+    ".vscode",
+    "__pypackages__",
+    "_build",
+    "buck-out",
+    "build",
+    "dist",
+    "node_modules",
+    "site-packages",
+    "venv",
+]
+line-length = 120
+indent-width = 4
+target-version = ['py312']
 
 [tool.ruff]
-line-length = 100
+line-length = 120
 target-version = "py312"
-extend-exclude = [".md"]
+extend-exclude = ["*.md"]
 
 [tool.ruff.lint]
 select = [
-    "F",  # pyflakes
-    "E", "W",  # pycodestyle
-    "A",  # flake8-builtins
-    "B",  # flake8-bugbear
-    "C4",  # flake8-comprehensions
-    "SIM",  # flake8-simplify
-    "ARG",  # flake8-unused-arguments
-    "PL",  # pylint
-    "TID",  # tidy
-    "I",  # isort
-    "N",  # pep8-naming
+    "F", # pyflakes
+    "E", "W", # pycodestyle
+    "A", # flake8-builtins
+    "B", # flake8-bugbear
+    "C4", # flake8-comprehensions
+    "SIM", # flake8-simplify
+    "ARG", # flake8-unused-arguments
+    "PL", # pylint
+    "TID", # tidy
+    "I", # isort
+    "N", # pep8-naming
     "UP"  # pyupgrade
 ]
 ignore = [
-    "E501",  # may cause conflicts with ruff formatter
+    "E501", # may cause conflicts with ruff formatter
     "W191"
 ]
 [tool.ruff.format]
